Moves CountryDealers to package country.dealers
