+++ core/src/main/countries/dealers/CountryDealer.kt
@@ -1,6 +1,7 @@
-package countries
+package countries.dealers
 
 import Player
+import countries.Occupation
 
 interface CountryDealer {
     fun dealTo(players: Collection<Player>): Collection<Occupation>
