+++ core/src/main/countries/dealers/CountryDealerExceptions.kt
@@ -1,4 +1,4 @@
-package countries
+package countries.dealers
 
 class NoPlayersToDealToException : Exception(
     "Can't deal to an empty collection of players.")
