+++ core/src/main/countries/dealers/RandomCountryDealer.kt
@@ -1,9 +1,9 @@
-package countries
+package countries.dealers
 
 import Country
 import Player
 
-class RandomCountryDealer(private val countries: List<Country>): CountryDealer {
+class RandomCountryDealer(private val countries: List<Country>) : CountryDealer {
     override fun dealTo(players: Collection<Player>) =
         DeterministicCountryDealer(countries.shuffled()).dealTo(players.shuffled())
 }
